# skyroster-backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: SkyRoster API
  description: Pilot schedule conflict detection system
  version: 0.1.0

security:
  - bearerAuth: []

paths:
  /api/health:
    get:
      operationId: getHealth
      tags: [Health]
      summary: Health check
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /api/me:
    get:
      operationId: getCurrentUser
      tags: [User]
      summary: Get current authenticated user info
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/compliance/reports:
    get:
      operationId: getComplianceReports
      tags: [Compliance]
      summary: List compliance reports
      responses:
        '200':
          description: List of compliance reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceReport'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/planning/schedules:
    get:
      operationId: getPlanningSchedules
      tags: [Planning]
      summary: List schedules
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/aircraft:
    post:
      operationId: addAircraft
      tags: [Aircraft]
      summary: Add a new aircraft to the fleet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAircraftRequest'
      responses:
        '201':
          description: Aircraft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftResponse'
        '400':
          description: Bad request (invalid type or base)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Aircraft with this registration already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getAircraft
      tags: [Aircraft]
      summary: List all aircraft in the fleet
      responses:
        '200':
          description: List of aircraft
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AircraftResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/aircraft/{id}:
    put:
      operationId: updateAircraft
      tags: [Aircraft]
      summary: Update an aircraft in the fleet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAircraftRequest'
      responses:
        '200':
          description: Aircraft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AircraftResponse'
        '400':
          description: Bad request (invalid type or base, or blank registration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aircraft not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Registration number is used by another aircraft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
        - timestamp
        - path
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 401
        error:
          type: string
          description: Error name
          example: "Unauthorized"
        message:
          type: string
          description: Human-readable error description
          example: "JWT token has expired"
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error occurrence
        path:
          type: string
          description: Request path
          example: "/api/admin/users"

    UserInfo:
      type: object
      required:
        - username
        - email
        - roles
      properties:
        username:
          type: string
          description: Username from JWT token
          example: "admin"
        email:
          type: string
          description: User email address
          example: "admin@skyroster.local"
        roles:
          type: array
          description: List of realm roles assigned to the user
          items:
            type: string
          example: ["operations_administrator"]

    HealthStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: Service health status
          example: "UP"

    ComplianceReport:
      type: object
      required:
        - id
        - title
        - status
      properties:
        id:
          type: string
          example: "CR-001"
        title:
          type: string
          example: "Monthly flight hours compliance"
        status:
          type: string
          example: "PASSED"

    Schedule:
      type: object
      required:
        - id
        - pilotName
        - date
      properties:
        id:
          type: string
          example: "SCH-001"
        pilotName:
          type: string
          example: "Jan Kowalski"
        date:
          type: string
          format: date
          example: "2026-03-20"

    AddAircraftRequest:
      type: object
      required:
        - registrationNumber
        - aircraftTypeCode
        - operationalBaseCode
      properties:
        registrationNumber:
          type: string
          description: Aircraft registration number
          example: "SP-LRA"
        aircraftTypeCode:
          type: string
          description: ICAO aircraft type code
          example: "B738"
        operationalBaseCode:
          type: string
          description: ICAO operational base code
          example: "EPWA"

    UpdateAircraftRequest:
      type: object
      required:
        - registrationNumber
        - aircraftTypeCode
        - operationalBaseCode
      properties:
        registrationNumber:
          type: string
          description: Aircraft registration number
          example: "SP-LRA"
        aircraftTypeCode:
          type: string
          description: ICAO aircraft type code
          example: "B738"
        operationalBaseCode:
          type: string
          description: ICAO operational base code
          example: "EPWA"

    AircraftTypeInfo:
      type: object
      required:
        - id
        - icaoCode
        - name
      properties:
        id:
          type: string
          format: uuid
        icaoCode:
          type: string
          example: "B738"
        name:
          type: string
          example: "Boeing 737-800"

    OperationalBaseInfo:
      type: object
      required:
        - id
        - icaoCode
        - name
      properties:
        id:
          type: string
          format: uuid
        icaoCode:
          type: string
          example: "EPWA"
        name:
          type: string
          example: "Warsaw Chopin"

    AircraftResponse:
      type: object
      required:
        - id
        - registrationNumber
        - aircraftType
        - operationalBase
      properties:
        id:
          type: string
          format: uuid
        registrationNumber:
          type: string
          example: "SP-LRA"
        aircraftType:
          $ref: '#/components/schemas/AircraftTypeInfo'
        operationalBase:
          $ref: '#/components/schemas/OperationalBaseInfo'